+++ Package.swift
@@ -5,7 +5,8 @@ let package = Package(
     name: "Murmur",
     platforms: [.macOS(.v26)],
     dependencies: [
-        .package(url: "https://github.com/argmaxinc/WhisperKit.git", from: "1.0.0")
+        .package(url: "https://github.com/argmaxinc/WhisperKit.git", from: "1.0.0"),
+        .package(url: "https://github.com/FluidInference/FluidAudio.git", from: "0.15.0"),
     ],
     targets: [
         .binaryTarget(
@@ -20,6 +21,7 @@ let package = Package(
             name: "Murmur",
             dependencies: [
                 .product(name: "WhisperKit", package: "WhisperKit"),
+                .product(name: "FluidAudio", package: "FluidAudio"),
                 "WhisperCppFramework",
                 "HarperFramework",
             ],